Use gemini prompt-interactive
